Fixup: Address review feedback
Ref: https://github.com/farmOS/farmOS-map/issues/68#issuecomment-836717251

Will be squashed in final mergeable pull-request version.
