+++ src/control/Edit/Edit.js
@@ -31,6 +31,7 @@ import './Edit.css';
  * @api
  */
 class Edit extends Control {
+
   /**
    * @param {Options=} opts Edit options.
    */
