+++ src/instance/instance.js
@@ -8,7 +8,7 @@ import defaults from './defaults';
 import addLayer, { getLayerByName } from './methods/layer';
 import addPopup from './methods/popup';
 import { zoomToVectors, zoomToLayer } from './methods/zoom';
-import { addBehavior, attachBehavior } from './methods/behavior';
+import { addBehavior, attachBehavior, attachBehaviorsByWeight } from './methods/behavior';
 import { measureGeometry } from '../utils/measure';
 
 // Define an object that represents a single farmOS map instance.
@@ -54,6 +54,7 @@ const createInstance = ({ target, options = {} }) => {
     zoomToLayer,
     addBehavior,
     attachBehavior,
+    attachBehaviorsByWeight,
     measureGeometry,
   };
